Restructured audio.rs
