+++ framework_crates/bones_framework/src/lib.rs
@@ -52,6 +52,9 @@ pub mod storage;
 pub mod time;
 pub mod utils;
 
+#[cfg(feature = "audio")]
+pub mod audio;
+
 #[cfg(feature = "ui")]
 pub mod debug;
 
@@ -93,7 +96,7 @@ impl lib::GamePlugin for DefaultGamePlugin {
     #[allow(unused_variables)]
     fn install(self, game: &mut lib::Game) {
         #[cfg(feature = "audio")]
-        game.install_plugin(render::audio::game_plugin);
+        game.install_plugin(audio::game_plugin);
 
         #[cfg(feature = "scripting")]
         game.install_plugin(bones_scripting::ScriptingGamePlugin::default());
